+++ source/sw/src/hornet.cpp
@@ -404,8 +404,8 @@ int DoHornetMatchPlayerZ(short SpriteNum)
         u->sz = bound;
     }
 
-    u->sz = min(u->sz, loz - u->floor_dist);
-    u->sz = max(u->sz, hiz + u->ceiling_dist);
+    u->sz = min(long(u->sz), long(loz - u->floor_dist));
+    u->sz = max(long(u->sz), long(hiz + u->ceiling_dist));
 
     u->Counter = (u->Counter + (ACTORMOVETICS << 3) + (ACTORMOVETICS << 1)) & 2047;
     sp->z = u->sz + ((HORNET_BOB_AMT * (long)sintable[u->Counter]) >> 14);
